完成

+++ algorithm/src/main/java/edu/gyj/bean/student/StudentToTalCS.java
@@ -5,13 +5,13 @@ import java.io.Serializable;
 public class StudentToTalCS implements Serializable {
     private String sid;
     private int consumption_count;
-    private float consumption_total_money;
-    private float consumption_average_money;
+    private double consumption_total_money;
+    private double consumption_average_money;
 
     public StudentToTalCS() {
     }
 
-    public StudentToTalCS(String sid, int consumption_count, float consumption_total_money, float consumption_average_money) {
+    public StudentToTalCS(String sid, int consumption_count, double consumption_total_money, double consumption_average_money) {
         this.sid = sid;
         this.consumption_count = consumption_count;
         this.consumption_total_money = consumption_total_money;
@@ -44,19 +44,19 @@ public class StudentToTalCS implements Serializable {
         this.consumption_count = consumption_count;
     }
 
-    public float getConsumption_total_money() {
+    public double getConsumption_total_money() {
         return consumption_total_money;
     }
 
-    public void setConsumption_total_money(float consumption_total_money) {
+    public void setConsumption_total_money(double consumption_total_money) {
         this.consumption_total_money = consumption_total_money;
     }
 
-    public float getConsumption_average_money() {
+    public double getConsumption_average_money() {
         return consumption_average_money;
     }
 
-    public void setConsumption_average_money(float consumption_average_money) {
+    public void setConsumption_average_money(double consumption_average_money) {
         this.consumption_average_money = consumption_average_money;
     }
 }
